Formtastic Bootstrap (how-to)
Flutie is gone (and who needs it anyway) but we can use the
formtastic-bootstrap gem to convert all our existing Formtastic
forms straightforwardly to Bootstrap markup.

There are also some required changes that will have to be
made to support the newer versions of Formtastic -- all those
f.buttons stanzas need to be changed to f.actions.